Update 3SI and Add VT Section
Shrinks the 3SI section and adds a new section next to it on wider screens for my time at Virginia Tech and what I gained from it. Because the sections are relatively small, they will display next to each other on larger screens. On smaller screens, they will display in a column.